new structure -> CKernel implements the hardware layer, CBootableKernel implements the func to start the kernel, CTacosKernel should implement the rest using the inherited members
