+++ src/kernel/Makefile
@@ -1,9 +1,13 @@
 export ROOT=$(CURDIR)/../../
+export CORE=$(ROOT)/src/kernel/core
 
 # ---  Target dependent and should be changed w.r.t architecture
-OBJS = main.o $(ROOT)/src/kernel/arch/rasp4/arch.o
+OBJS = main.o \
+		$(CORE)/boot.o \
+		$(CORE)/tacoskernel.o \
+		$(ROOT)/src/kernel/arch/rasp4/arch.o
 
-LIBS	= $(ROOT)/src/kernel/core/core.a $(CIRCLEHOME)/lib/libcircle.a
+LIBS	= $(CIRCLEHOME)/lib/libcircle.a
 
 include $(ROOT)/src/misc/common-pi4.mk
 include $(ROOT)/src/misc/common-rules.mk
@@ -11,7 +15,6 @@ include $(ROOT)/src/misc/common-rules.mk
 # ---
 .deps: 
 	@echo "deps"
-	$(MAKE) -C $(CURDIR)/core
 	cd $(ROOT)/aarch/rasp4 && ./makeall
 
 .PHONY: all clean deps
