--- src/kernel/arch/rasp4/arch.h
@@ -1,32 +1,0 @@
-#ifndef _kernel_h
-#define _kernel_h
-
-#include <circle/memory.h>
-#include <circle/actled.h>
-#include <circle/types.h>
-
-enum TShutdownMode
-{
-	ShutdownNone,
-	ShutdownHalt,
-	ShutdownReboot
-};
-
-class CKernel
-{
-public:
-	CKernel (void);
-	~CKernel (void);
-
-	boolean Initialize (void);
-
-	TShutdownMode Run (void);
-
-private:
-	// do not change this order
-	CMemorySystem	m_Memory;
-	CActLED		m_ActLED;
-	
-};
-
-#endif
