+++ src/kernel/core/boot.cpp
@@ -1,20 +1,7 @@
-#include <arch.h>
 #include <errcode.h>
-// int main()
-// {
-//     /* code */
-//     return 0;
-// }
+#include "tacoskernel.h"
 
 ret_t boot(){
-    return init_arch();
+    CTacosKernel kernel;
+    return kernel.Start();
 }
-
-// void freeTACOS()
-// {
-//     freeTACOS_arch_init();
-//     freeTACOS_bsp_init();
-
-//     main();
-// }
-
