+++ src/kernel/core/tacoskernel.cpp
@@ -1,0 +1,9 @@
+#include "tacoskernel.h"
+
+CTacosKernel::CTacosKernel()
+{
+}
+
+CTacosKernel::~CTacosKernel()
+{
+}
