+++ src/kernel/core/tacoskernel.h
@@ -1,0 +1,10 @@
+#include <arch.h>
+
+class CTacosKernel : public CBootableKernel
+{
+private:
+    
+public:
+    CTacosKernel();
+    ~CTacosKernel();
+};
