--- src/kernel/include/bsp.h
@@ -1,8 +1,0 @@
-#ifndef __FREETACOS_BSP__
-#define __FREETACOS_BSP__
-
-#include <errno.h>
-
-// freeTACOS_ret_t freeTACOS_bsp_init();
-
-#endif
