+++ src/kernel/include/core/boot.h
@@ -1,6 +1,8 @@
 #ifndef __FREETACOS_BOOT__
 #define __FREETACOS_BOOT__
 
+#include <errcode.h>
+
 void freeTACOS_boot();
 
 ret_t boot();
