--- src/kernel/include/core/error.h
@@ -1,4 +1,0 @@
-#ifndef __FREETACOS_ERROR__
-#define __FREETACOS_ERROR__
-
-#endif
