--- src/kernel/include/core/partition.h
@@ -1,4 +1,0 @@
-#ifndef __FREETACOS_PARTITION__
-#define __FREETACOS_PARTITION__
-
-#endif
