--- src/kernel/include/core/timer.h
@@ -1,4 +1,0 @@
-#ifndef __FREETACOS_TIMER__
-#define __FREETACOS_TIMER__
-
-#endif
